backend/requirements.txt
# Ядро
mediapipe==0.10.9
protobuf==3.20.3

# Математика и Видео
opencv-python==4.8.0.76
numpy==1.26.3

pyzmq